+++ efrei-internship-front/pages/about.vue
@@ -3,14 +3,6 @@
 </script>
 
 <template>
-    <div>
-        <header>
-            <nav>
-                <ul>
-                    <li><NuxtLink to="/">Home</NuxtLink></li>
-                </ul>
-            </nav>
-        </header>
-        <h1>About page</h1>
-    </div>
+    <HeaderETU />
+    <h1>About page</h1>
 </template>
