+++ efrei-internship-front/pages/index.vue
@@ -1,22 +1,24 @@
 <script setup lang="ts">
 import { useSessionStore } from '#imports';
-const sessionStore = useSessionStore();
+// const sessionStore = useSessionStore();
 
-const userInfos = ref(sessionStore.getUser)
+// const userInfos = ref(sessionStore.getUser)
 
-const logout = () => {
-  useFetch('http://localhost:3000/api/logout', {
-    method: 'POST',
-  }).then(() => {
-    sessionStore.logout()
-    navigateTo('/login')
-  })
-}
+// const logout = () => {
+//   useFetch('http://localhost:3000/api/logout', {
+//     method: 'POST',
+//   }).then(() => {
+//     sessionStore.logout()
+//     navigateTo('/login')
+//   })
+// }
 
 </script>
 
 <template>
-  <div class="m-5">
+  <HeaderETU />
+  <div class="text-center mt-2">This is the index page of the student Front-End</div>
+  <!-- <div class="m-5">
     <header class="flex flex-row justify-between">
       <nav>
         <ul class="text-blue-600">
@@ -44,5 +46,5 @@ const logout = () => {
     </header>
     <div class="text-2xl text-center">Home Page - Child component below</div>
     <ButtonCounter />
-  </div>
+  </div> -->
 </template>
